Refactoring #1
