--- frontend/src/components/Workflow/Sidebar.tsx
@@ -1,25 +1,0 @@
-import { useDnD } from './DnDContext';
-
-export default () => {
-    const [_, setType] = useDnD();
-
-    const onDragStart = (event, nodeType: string) => {
-        setType(nodeType);
-        event.dataTransfer.effectAllowed = 'move';
-    };
-
-    return (
-        <aside>
-            <div className="description">You can drag these nodes to the pane on the right.</div>
-            <div className="dndnode input" onDragStart={(event) => onDragStart(event, 'input')} draggable>
-                Input Node
-            </div>
-            <div className="dndnode" onDragStart={(event) => onDragStart(event, 'default')} draggable>
-                Default Node
-            </div>
-            <div className="dndnode output" onDragStart={(event) => onDragStart(event, 'output')} draggable>
-                Output Node
-            </div>
-        </aside>
-    );
-};
